[Editor] Moving image upload to useQuill
