+++ components/Comment/CommentEditor.tsx
@@ -33,8 +33,6 @@ import { isEmpty as isInputEmpty } from "~/config/utils/nullchecks";
 import ResearchCoinIcon from "../Icons/ResearchCoinIcon";
 import Bounty from "~/config/types/bounty";
 import { ModalActions } from "~/redux/modals";
-import api from "~/config/api";
-import { Helpers } from "@quantfive/js-web-config";
 
 const { setMessage, showMessage } = MessageActions;
 
@@ -85,48 +83,7 @@ const CommentEditor = ({
     commentType || commentTypes.find((t) => t.isDefault)!.value
   );
 
-  const getFileUrl = ({ fileString, type }) => {
-    const payload = {
-      content_type: type.split("/")[1],
-      content: fileString,
-    };
-    return fetch(api.SAVE_IMAGE, api.POST_CONFIG(payload))
-      .then(Helpers.checkStatus)
-      .then(Helpers.parseJSON)
-      .then((res) => {
-        return res;
-      });
-  };
 
-  const toBase64 = (file) =>
-    new Promise((resolve, reject) => {
-      const reader = new FileReader();
-      reader.readAsDataURL(file);
-      reader.onload = () => resolve(reader.result);
-      reader.onerror = (error) => reject(error);
-    });
-
-  const imageHandler = (obj) => {
-    const input = document.createElement("input");
-    input.setAttribute("type", "file");
-    input.setAttribute("accept", "image/*");
-    input.click();
-    input.onchange = async function () {
-      // showLoader(true);
-      const file = input.files[0];
-      const fileString = await toBase64(file);
-      const type = file.type;
-      const fileUrl = await getFileUrl({ fileString, type });
-      if (quill) {
-        const range = quill.getSelection();
-
-        // this part the image is inserted
-        // by 'image' option below, you just have to put src(link) of img here.
-        quill.insertEmbed(range.index, "image", fileUrl);
-      }
-      // showLoader(false);
-    };
-  };
 
   const { quill, quillRef, isReady } = useQuill({
     options: {
@@ -142,13 +99,6 @@ const CommentEditor = ({
     content,
     notifyOnContentChangeRate: 300, // ms
   });
-
-  useEffect(() => {
-    if (quill) {
-      // Add custom handler for Image Upload
-      quill.getModule("toolbar").addHandler("image", imageHandler);
-    }
-  }, [quill]);
 
   useEffectForCommentTypeChange({
     quill,
